Reorganize codebase by domain and add comprehensive documentation
- Restructure src/ into domain-based modules:
  - core/: RoomRoutine base class and lifecycle management
  - routines/: Bootstrap, EnergyMining, EnergyCarrying, Construction
  - spatial/: RoomMap with peak detection and territory division
  - planning/: GOAP agent framework
  - types/: Domain interfaces (SourceMine, EnergyRoute, Memory)
  - utils/: ErrorMapper and helpers

- Add comprehensive documentation:
  - README.md: Project overview, quick start, architecture
  - docs/ARCHITECTURE.md: System design and module structure
  - docs/ECONOMIC_FRAMEWORK.md: ROI tracking and market mechanics
  - docs/ROUTINES.md: Colony operations guide
  - docs/SPATIAL_SYSTEM.md: RoomMap algorithms

- Add detailed JSDoc comments throughout codebase
- Update all imports to use new module paths
- Export modules through index.ts barrel files
