--- src/EnergyRoute.ts
@@ -1,4 +1,0 @@
-export interface EnergyRoute {
-    waypoints : {x: number, y: number, roomName: string, surplus: boolean}[];
-    Carriers : { creepId: Id<Creep>, waypointIdx: number }[];
-}
