--- src/SourceMine.ts
@@ -1,6 +1,0 @@
-export interface SourceMine {
-    sourceId : Id<Source>;
-    HarvestPositions: RoomPosition[];
-    flow: number;
-    distanceToSpawn: number;
-}
