+++ src/core/index.ts
@@ -1,0 +1,14 @@
+/**
+ * @fileoverview Core module exports.
+ *
+ * This module provides the foundational classes and interfaces
+ * for the colony routine system.
+ *
+ * @module core
+ */
+
+export {
+  RoomRoutine,
+  ResourceContract,
+  PerformanceRecord,
+} from "./RoomRoutine";
