--- src/custom.d.ts
@@ -1,9 +1,0 @@
-interface RoomMemory {
-    routines: {
-        [routineType: string]: any[];
-    };
-}
-
-interface CreepMemory {
-    role?: string;
-}
