+++ src/planning/index.ts
@@ -1,0 +1,18 @@
+/**
+ * @fileoverview Planning module exports.
+ *
+ * This module provides GOAP (Goal-Oriented Action Planning) components
+ * for agent behavior planning.
+ *
+ * @module planning
+ */
+
+export {
+  Action,
+  Goal,
+  WorldState,
+  Agent,
+  createMineEnergyAction,
+  createBuildStructureAction,
+  createProfitGoal,
+} from "./Agent";
