+++ src/routines/index.ts
@@ -1,0 +1,16 @@
+/**
+ * @fileoverview Colony routines module exports.
+ *
+ * This module provides all colony operation routines:
+ * - Bootstrap: Early-game colony initialization
+ * - EnergyMining: Harvester management at sources
+ * - EnergyCarrying: Logistics and resource transport
+ * - Construction: Builder management for construction sites
+ *
+ * @module routines
+ */
+
+export { Bootstrap } from "./Bootstrap";
+export { EnergyMining } from "./EnergyMining";
+export { EnergyCarrying } from "./EnergyCarrying";
+export { Construction } from "./Construction";
