+++ src/spatial/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @fileoverview Spatial analysis module exports.
+ *
+ * This module provides spatial analysis tools for colony planning:
+ * - RoomMap: Full room analysis with peaks and territories
+ * - Peak: Optimal building location interface
+ * - Territory: Zone ownership interface
+ *
+ * @module spatial
+ */
+
+export { RoomMap, Peak, Territory } from "./RoomMap";
