+++ src/types/index.ts
@@ -1,0 +1,14 @@
+/**
+ * @fileoverview Domain types module exports.
+ *
+ * This module provides all domain-specific type definitions
+ * used throughout the colony system.
+ *
+ * @module types
+ */
+
+export { SourceMine } from "./SourceMine";
+export { EnergyRoute, RouteWaypoint, CarrierAssignment } from "./EnergyRoute";
+
+// Memory types are global declarations, not exports
+// See Memory.ts for Screeps memory extensions
