+++ src/utils/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @fileoverview Utilities module exports.
+ *
+ * This module provides utility functions and helpers.
+ *
+ * @module utils
+ */
+
+export { ErrorMapper } from "./ErrorMapper";
